Move Mock Server into it's own file
